pyproject.toml
[project]
name = "happi"
version = "0.1.0"
description = "Turn any OpenAPI spec into a human-friendly CLI"
readme = "README.md"
authors = [
    { name = "Elad Ben-Haim", email = "elad12390@gmail.com" }
]
requires-python = ">=3.12"
license = "Apache-2.0"
keywords = ["openapi", "cli", "api", "rest", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.24.0",
    "rich>=14.0.0",
    "httpx>=0.28.0",
    "prance>=23.6.0",
    "openapi-spec-validator>=0.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "inflect>=7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pyright>=1.1.0",
    "pytest>=8.0",
    "pytest-bdd>=8.0",
    "pytest-asyncio>=0.25.0",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.0",
]

[project.scripts]
happi = "happi.cli:app_entry"

[project.urls]
Homepage = "https://github.com/elad12390/happi"
Repository = "https://github.com/elad12390/happi"
Issues = "https://github.com/elad12390/happi/issues"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "TCH",
    "RUF",
    "PTH",
    "ERA",
    "TID",
    "PIE",
    "T20",
    "S",
    "C4",
    "DTZ",
    "FA",
    "ISC",
    "PGH",
    "RSE",
    "RET",
    "FLY",
    "PERF",
]
ignore = [
    "S101",
]

[tool.ruff.lint.per-file-ignores]
".bdd/**" = ["T20", "S"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
include = ["src"]
